bin: move common functions,structs to cli.rs
