+++ Cargo.toml
@@ -10,6 +10,10 @@ documentation = "https://docs.rs/elasticsearch_exporter"
 homepage = "https://github.com/vinted/elasticsearch_exporter.rs"
 license = "BSD-3-Clause"
 repository = "https://github.com/vinted/elasticsearch_exporter.rs"
+autobins = false
+
+[[bin]]
+name = "elasticsearch_exporter"
 
 [dependencies]
 elasticsearch = "7.10.0-alpha.1"
